requirements.txt
langchain==0.1.13
langchain-core==0.1.33
langgraph==0.0.45
langchain-google-genai==0.0.10
google-generativeai==0.3.5
python-dotenv==1.0.0
pydantic==2.5.0
